backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# Rubik's Cube Solver
kociemba==1.2.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Additional utilities
python-dateutil==2.8.2

# ML/AI (optional - for advanced detection)
# tensorflow==2.15.0
# torch==2.1.2
# torchvision==0.16.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0